ibc_types::core::channel::events
Module packet

Structs

AcknowledgePacket

ChannelClose

A ChannelClose event is emitted when a channel is closed as a result of a packet timing out. Note that since optimistic packet sends (i.e. send a packet before channel handshake is complete) are supported, we might not have a counterparty channel id value yet. This would happen if a packet is sent right after a ChannelOpenInit message.

ReceivePacket

SendPacket

TimeoutPacket

WriteAcknowledgement