ibc_types::core::channel
Module msgs

Message definitions for all ICS4 domain types: channel open & close handshake datagrams, as well as packets.

Structs

MsgAcknowledgement

Message definition for packet acknowledgements.

MsgChannelCloseConfirm

Message definition for the second step in the channel close handshake (the ChanCloseConfirm datagram). Per our convention, this message is sent to chain B.

MsgChannelCloseInit

Message definition for the first step in the channel close handshake (ChanCloseInit datagram). Per our convention, this message is sent to chain A.

MsgChannelOpenAck

Per our convention, this message is sent to chain A. Message definition for the third step in the channel open handshake (ChanOpenAck datagram).

MsgChannelOpenConfirm

Message definition for the fourth step in the channel open handshake (ChanOpenConfirm datagram). Per our convention, this message is sent to chain B.

MsgChannelOpenInit

Message definition for the first step in the channel open handshake (ChanOpenInit datagram). Per our convention, this message is sent to chain A.

MsgChannelOpenTry

Message definition for the second step in the channel open handshake (ChanOpenTry datagram). Per our convention, this message is sent to chain B.

MsgRecvPacket

Message definition for the “packet receiving” datagram.

MsgTimeout

Message definition for packet timeout domain type, which is sent on chain A and needs to prove that a previously sent packet was not received on chain B

MsgTimeoutOnClose

Message definition for packet timeout domain type.

Enums

ChannelMsg

Enumeration of all possible messages that the ICS4 protocol processes.

PacketMsg