ibc_types::core::connection
Module events

Types for ABCI Events that inform relayers about IBC client events.

Structs

ConnectionOpenAck

Per our convention, this event is generated on chain A.

ConnectionOpenConfirm

Per our convention, this event is generated on chain B.

ConnectionOpenInit

Per our convention, this event is generated on chain A.

ConnectionOpenTry

Per our convention, this event is generated on chain B.

Enums

Error

An error while parsing an Event.