ibc_types
Trait DomainType

pub trait DomainType:
    Sized
    + Clone
    + TryFrom<Self::Proto>
where
    Self::Error: Into<Error> + Send + Sync + 'static,
{
    type Proto: Message + Name + Default + From<Self> + Send + Sync + 'static;

    // Provided methods
    fn encode_to_vec(&self) -> Vec<u8> { ... }
    fn to_proto(&self) -> Self::Proto { ... }
    fn decode<B>(buf: B) -> Result<Self, Error>
       where B: Buf { ... }
}

A marker type that captures the relationships between a domain type (Self) and a protobuf type (Self::Proto).

Required Associated Types

type Proto: Message + Name + Default + From<Self> + Send + Sync + 'static

Provided Methods

fn encode_to_vec(&self) -> Vec<u8>

Encode this domain type to a byte vector, via proto type P.

fn to_proto(&self) -> Self::Proto

Convert this domain type to the associated proto type.

This uses the From impl internally, so it works exactly like .into(), but does not require type inference.

fn decode<B>(buf: B) -> Result<Self, Error>
where B: Buf,

Decode this domain type from a byte buffer, via proto type P.

Object Safety

This trait is not object safe.